pyproject.toml
[tool.poetry]
name = "personality-bfi"
version = "0.1.0"
description = "Big Five Inventory Personality test (French version) in a python web application"
license = ""
authors = ["Paul Dulaud <paul.dulaud@utt.fr>"]
readme = "README.rst"
homepage = "https://gitlab.com/phd-connected-objects-for-proactive-health-management/personality/big-five-inventory-python-web-app"
repository = "https://gitlab.com/phd-connected-objects-for-proactive-health-management/personality/big-five-inventory-python-web-app"

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
Flask = "^1.1.2"
pandas = "^1.2.0"
Flask-WTF = "^0.14.3"
Flask-Assets = "^2.0"
lesscpy = "^0.14.0"
cssmin = "^0.2.0"
jsmin = "^2.2.2"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
